updating amenity for error handling od none

+++ api/v1/views/amenities.py
@@ -33,9 +33,12 @@ def get_a_amenity(amenity_id):
                  strict_slashes=False)
 def delete_amenity(amenity_id):
     """ Deletes a Amenity object """
+    amenity = storage.get("Amenity", amenity_id)
+    if amenity is None:
+        abort(404)
     try:
-        amenity = storage.get("Amenity", amenity_id)
         storage.delete(amenity)
+        storage.save()
         return jsonify({}), 200
     except:
         abort(404)
@@ -44,29 +47,32 @@ def delete_amenity(amenity_id):
 @app_views.route('/amenities', methods=['POST'], strict_slashes=False)
 def create_amenity():
     """ Creates a Amenity """
-    if not request.get_json():
+    try:
+        content = request.get_json()
+        if 'name' not in content:
+            return "Missing name", 400
+        amenity = Amenity(content)
+        amenity.save()
+        new_amenity = storage.get("Amenity", amenity.id)
+        return jsonify(new_amenity.to_json()), 201
+    except:
         return "Not a JSON", 400
-    if not 'name' in request.get_json():
-        return "Missing name", 400
-    amenity = Amenity(request.get_json())
-    amenity.save()
-    new_amenity = storage.get("Amenity", amenity.id)
-    return jsonify(new_amenity.to_json()), 201
 
 
 @app_views.route('/amenities/<amenity_id>', methods=['PUT'],
                  strict_slashes=False)
 def update_amenity(amenity_id):
     """ Updates a Amenity """
-    if not request.get_json():
-        return "Not a JSON", 400
+    amenity = storage.get("Amenity", amenity_id)
+    if amenity is None:
+        abort(404)
     try:
-        amenity = storage.get("Amenity", amenity_id).to_json()
         skip_list = ["id", "created_at", "updated_at"]
         key_values = request.get_json()
+        amenity = amenity.to_json()
         for k, v in key_values.items():
             if k not in skip_list:
                 amenity[k] = v
         return jsonify(amenity), 200
     except:
-        abort(404)
+        return "Not a JSON", 400
